requirements.txt
# Core AI/ML dependencies
torch>=1.13.0
transformers>=4.25.0
sentence-transformers>=2.2.0
llama-index>=0.9.0
faiss-cpu>=1.7.4

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0
pandas>=1.5.0

# Utilities
numpy>=1.21.0
markdown>=3.4.0

# Optional: For better performance
# accelerate>=0.20.0
# bitsandbytes>=0.39.0